kex.c: fix compiler warning
```
src/kex.c:4046:..: warning C4244: 'initializing': conversion from '__int64' to 'int', possible loss of data
```

+++ src/kex.c
@@ -4042,7 +4042,7 @@ libssh2_session_method_pref(LIBSSH2_SESSION * session, int method_type,
 
     while(s && *s && mlist) {
         char *p = strchr(s, ',');
-        int method_len = p ? (p - s) : (int) strlen(s);
+        size_t method_len = (p ? (size_t)(p - s) : strlen(s));
 
         if(!kex_get_method_by_name(s, method_len, mlist)) {
             /* Strip out unsupported method */
